requirements.txt
docx2pdf==0.1.8
Flask==3.0.2
mysql_connector_repackaged==0.3.1
supabase==2.4.0
Werkzeug==3.0.1
gunicorn
